backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
python-decouple==3.8
psycopg2-binary==2.9.9
dj_database_url==1.0.0
anthropic==0.23.1
django-cors-headers==4.3.1
